node_operations_fuzzy_test: repro some issues
This patch separates out the workload generator and runner from the test
body. This is useful in helping separate out concerns (maybe one day we
can pull a NodeOpFuzzer out of the test), and for allowing us to reuse
the code to exercise specific sequences.

There are two bad sequences that seem legal, but actually end up
failing.