utils: add UUID type
This commit adds a new type uuid_t that will soon be used for node UUID
and cluster UUID. It wraps boost::uuids::uuid and includes some hooks
required for serialization and hashing.

Note that already there exists a kafka::uuid type that wraps
`std::array<uint18_t, 16>`. I opted to avoid reusing it since it is
exposed via the Kafka API, making it difficult to modify in the future
if need be.