rpc: add simple_protocol server
As a part of some cluster bootstrap reworking, it will be useful to
start a subset of the RPC services up before any others are started. To
that end, it'd be nice if there were a way to add services to an RPC
server.

This commit adds a new wrapper for a server of the rpc::simple_protocol
that provides an interface for adding services to the protocol while the
server is running. Since the protocol used by net::server is generic,
this uses a cast into a rpc::simple_protocol to call into this new
RPC-specific functionality. While I'm generally not a fan of casting,
its scope in this case is fairly restricted, and it seems difficult to
misuse the new API.

I considered an alternate approach where we swap in a new protocol to
the server, but it was difficult to reconcile the destruction of the old
protocol being used by the running server. I also considered extending
this behavior generically for all server/protocols, but, as services are
an RPC-layer abstraction (i.e. not Kafka, etc), the new functionality
seems custom to an RPC-specific server.