raft: handle failed heartbeat
We weren't handling the case where a reply was seen but the request
failed, e.g. because the request was sent to the wrong node ID.

In practice it seems very difficult to make this happen because Redpanda
currently forbids new nodes from joining the cluster if another node
already exists on that server. This will need to change if we want to
support wiping and restarting a node without decommissioning.