cluster_discovery: read node ID from kvstore
When the controller starts up, we check to ensure the configured node ID
is identical to that in the kvstore if one exists, persisting one if
not. We should use this node ID at startup if weren't configured with
one, e.g. if we're upgrading an existing cluster and begin omitting node
ID from the node config.