storage: check for file before reading during compaction
It's possible that when merging adjacent segments, we pick a segment
that doesn't have a compaction index yet. In doing so, we previously
created the file when trying to read it, leading us to leave the file
behind.

This commit checks or the file before attempting to read it.

Fixes #8153