config: change default segment size to 128MiB
This size is a more suitable default, to have
finer granularity of disk usage, and finer
granularity of tiered storage read cache when
tiered storage is in use.

For systems that write high bandwidths to a single
partition, this will increase the number of segment
rolls, but should not outright break anything unless
the system was already very close to limits of e.g.
file handles or memory.

Fixes https://github.com/redpanda-data/redpanda/issues/8170