cluster: update signature of persisted_stm::wait_no_throw
There's an upcoming change to remove the timeout from the
archival_metadata_stm. Making that change will be much simpler if the
signature of persisted_stm::wait_no_throw() is a timeout instead of a
duration, since the underlying raft::state_machine::wait() call uses the
model::no_timeout deadline value to determine whether to set a timer.

Along the way I added an optional abort source reference that is
currently unused.