tests: add utilities for producing and consuming
In planning some new tests with tiered storage, I've been wanting to
write single-node unit tests that exercise produce and consume paths
from the Kafka layer down to underlying storage. It is particularly
important for tiered storage to test this bridge since the Kafka layer
is where we make key decisions regarding what storage subsystems to
direct to.

This commit adds some utilities to facilitate this: it wraps the
kafka::client::transport and abstracts away details of Kafka requests,
instead exposing a simple key-value interface for producing and
consuming.