cloud_metadata: add uploader class
This commit adds the structure for an upload loop that will upload
cluster metadata. The longer-term goal here will be to have this loop
upload things like a controller snapshot, but for now it just uploads a
manifest with an increasing ID.

It currently isn't plugged in anywhere, but the intent will be to
instantiate an uploader as a part of the controller.

Unit tests are added to exercise the basic utilities included
(downloading and uploading the manifest).