cloud_metadata: make new uploader loop use a high metadata ID
If starting an upload after the demise of a cluster, we need to be able
to differentiate between metadata from an old cluster and a new cluster.
A recovery process must be able to pick the metadata from the new
cluster upon restoring.

To that end, the metadata ID used by clusters restored from the same
backet need to have monotonically increasing metadata IDs.