cluster: pull out to-ntp-leader dispatch from id_allocator_frontend
This introduces a leader_routing_frontend class that encapsulates the
bits of id_allocator_frontend that find an NTP leader and route to the
shard that hosts it.

This will be useful for an upcoming change I have to target leaders of
the offsets topic, as well as introducing a new call into the
id_allocator to reset the next id.