cloud_storage: add kafka start offset to timequery params
This commit is a piece in an on-going effort to decouple the
start_kafka_offset override from the start offset maintained by cloud
storage. The overarching idea will be to consider the override only
where needed (during fetches and when applying retention), and otherwise
ignore the override.

To that end, this commit updates the fetch path to consider the
override: a timequery has an inherent lower bound that is the start of
the partition. This was previously represented as query parameters that
include a timestamp and a start offset, but not represented in the
async_manifest_view's querying interface, where it only accepts either a
timestamp _or_ an offset.

This updates the timestamp-based querying to accept a new
timequery_params struct that includes the Kafka start offset.