cloud_storage: seek forward if timequery falls below start override
When finding the right segment to scan in a given manifest, we don't
currently honor both search parameters (timestamp + kafka offset),
favoring timestamps always if they exist.

This updates the behavior to also seek forward to the kafka offset if
after querying by timestamp, the kafka offset is higher. This will be
useful in forcing the fetch path to honor the start_kafka_offset
override.