storage: fix comparison against last record in segment
We were previously comparing a given offset with the segment's last
offset, which in theory is the last record's offset, but across
compactions across multiple nodes, may end up not corresponding to a
real record. Take for example the following sequence:

node 1:
- Appends in order, end up with a segment [1707, 2000]. This segment contains
  batch [1785, 1799] with 15 records.
- Compaction runs, batch [1785, 1799] maintains its batch boundaries,
  but now has 6 records, e.g. the last one being 1790.
- The batches are replicated to node 2.

node 2:
- We have a different segment layout, but the batch ends up in segment
  [1230, 1799], still with only 6 records, the last one being 1790.
- We compact, and try to keep offset 1799, since this supposed to be the
  last offset.
- The last offset in the pre-compacted segment is actually 1790, and we
  end up not keeping the last batch because of this faulty comparison.
- The last batch in the post-compacted segment is actually 1784, and
  that's what is used for the max offset of the new segment.
- There is now a gap between the new compacted segment ending at 1784
  and the next, starting at 1800.

So, rather than comparing each record's offset with the segment's last
offset, this commit updates the check to compare the number of records
processed if the batch is the last batch in the segment.