storage: fix compaction short circuit check, redux
The previous short-circuit check for compaction aimed to compare the
first compactible offset against the max offset, which would supposedly
tell us whether the only data record were the last record, indicating a
sole data record that wouldn't be compacted.

This check didn't take into account batches of size > 1, whose first
data batch base offset would never equal the max offset (a batch end
offset).

This commit fixes this by changing this field to count the number of
compactible records appended. This value is sometimes an overestimate,
e.g. in cases where the segment is truncated and we don't count the
number of compactible records removed to subtract. But, that should be
sufficient since our check for having compactible records is okay with
false positives.