cluster/tests: add test case for eviction of the tx coordinator
CONFLICT:
- different CMakeLists surroundings
- different accessor for tm_stm instead of stm_manager

We've seen in older versions of Redpanda that despite the 'delete'
cleanup policy being enabled, eviction may not be triggered. This is
because in older versions of Redpanda[1], we would not register the
eviction STM for tx coordinator partitions.

[1] https://github.com/redpanda-data/redpanda/blob/3ce012f9ccb64810eafcc4b8b2e0df7340172879/src/v/cluster/partition.cc#L446-L471

(cherry picked from commit 761dfe987b3e4a97016ecbe966cc55c432aab3df)