cloud_storage: use path provider for spillovers in manifest view
The manifest view transparently downloads and caches spillover manifests
when Redpanda reads in the spillover region. This commit updates the
relevant path generation code to use the path provider.

The cursor doesn't need the path provider for anything but logging, so
instead of logging the full paths, this just has the cursor log the
filenames.