cloud_storage: use partition_manifest_downloader in scrubber/anomaly detector
The anomaly detector (and by extension, the scubber)  needs to be able
to find a partition manifest whose format (serde or json) is not yet
known. This is handled transparently by the partition manifest
downlaoder, while also taking remote labels into account.

This commit updates anomaly detector to use the downloader, and plugs in
the archiver's path provider from the scrubber.

As a part of this move, there's an anomalous case that is removed: when
we detect a JSON manifest that's a spillover manifest. Redpanda never
published JSON spillover manifests, so indeed this would be anomalous.
In practice this should never be the case, and this commit just removes
it, since it depends on knowing the downloaded format, which is no
longer the case when using the downloader.