iceberg: add manifest struct
Adds an implementation of the Apache Iceberg manifest, which represents
a set of files to be added or removed from a table.

Note, in this implementation we represent the manifest entries (i.e. the
data files) using structs generated using avrogencpp. These structs are
restrictive in a few ways:
- there is no built-in comparator
- comparators are difficult to implement since the generated code makes
  heavy use of std::any in Avro union types
- the manifest_entry r102 ("partition"[1]) field is incomplete, as it depends
  on runtime partitioning and can't be statically generated

I've left a TODO to address this, but it's still worth checking this
code in to allow for testing of other pieces of manifest serialization.

[1] https://iceberg.apache.org/spec/#manifests