iceberg: add field type copy methods
When applying transforms to a record to get the Iceberg
partition key, we'll also need to track the corresponding transformed
types and serialize them.

In order to transform some types, a copy method is required, e.g.
because the identity-transformed type would be a copy of the type.

We won't necessarily need deep copies immediately, but it seems
worthwhile to have utilities to make copies if they do become needed.
So, this commit adds a copy method for type fields.