iceberg: ::schema to avro::Schema
Adds methods for converting an iceberg::nested_field to an avro::Schema.
Our immediate need is to be able to write Iceberg manifest_entries,
which have a dynamic schema based on the partition key, but this code
will likely be useful when we begin translating Avro data to Parquet.

There are a few quirks of our usage of Avro to call out:
- Field IDs must be included as a custom attribute to each field,
  specified by the Icberg spec.
- Avro map types and record types in the manifest entries abide by the
  naming convention where records are named "r<field-id>" and maps are
  named "k<key-id>_v<val_id>". This is encoded in these methods, though
  may not be the right place for them longer term, if this code is
  reused for Avro data translation.
- avro::FixedSchema takes an input argument name field that doesn't
  appear to be useful, so empty strings are used instead.