iceberg: add struct accessors
This adds an accessor to child values of a struct value, to be used when
building a partition key for a given struct value. Since partition keys
can only be constructed from fields that are not a part of lists or
maps, the accessors in this commit do not touch lists or maps.

The implementation resembles the Iceberg Rust implementation[1].

[1] https://github.com/apache/iceberg-rust/blob/70d5b0b49e6e18676da82434ed1ee9a68d00d34b/crates/iceberg/src/expr/accessor.rs#L26-L85