iceberg: add id filter to schema type map builder
Improves on the id-type map-building by only looking for a subset of
fields. Given the use case for the map is to find and transforms the
types specified by the partition_spec (which at least starting out, will
only be a single field), building the full map of types is wasteful.