schema: add method for getting valid_schemas
Adds a method for getting ppsr::valid_schemas, so that the
schema::registry can implement. This will be useful for Iceberg schema
resolution since we'll need to operate on the full fidelity schemas in
their respective types, which the valid_schema contains.

As a part of this change, this added a getter interface that
schema::registry and the fake_registry now implement, to allow reuse of
the code the gets the valid schema.