cluster_restore_test: make spark lookup concurrent
This test uses a function that queries each table to look for the highest
offset in each table. This lookup was done serially, and meant that the
overarching function could take a couple of minutes. This resulted in
test timeouts when waiting for the translated offsets to reach a certain
point.

This commit parallelizes the function.