ct/l1: introduce domain manager
Introduces the domain_manager, an abstraction that will manage access to
the L1 domains, each backed by a partition of the L1 metastore topic.
To this end, each domain manager wraps an STM owned by the partition.

The STM used currently is the simple_stm. This allows the the domain
manager to reuse some code from simple_metastore to serve requests, but
with a state backed by a Raft group rather than one entirely in memory.

The direct caller of the domain manager will be the frontend, as the
endpoint of RPC calls. As such, the domain manager exposes of interface
with rpc:: request/reply types.