ct/l1: unify L1 topic creation in domain supervisor
Previously the frontend and the domain_supervisor each had roughly
equivalent code to create the L1 metastore topic. Now that the frontend
has access to the domain supervisor, this PR updates it to call into the
supervisor to create the metastore topic.

Along the way, this updates the behavior to create more than one
partition, as was done by the frontend previously.