ct/l1/lsm: add STM
Introduces a new STM that manages an instance of the newly introduced
lsm_state. This is meant to be used along with object storage to manage
a single lsm::database across replicas.

An upcoming change will introduce a new lsm::metadata_persistence that
will replicate LSM manifests using this newly introduced STM so that
lsm::databases can be opened from this STM on leaders.